test(cypress): fix Move mountComponentWithAppShell method from component.js to commands-components.js
